Remove the WTypes2.LPWSTRByReference class
The WTypes2.LPWSTRByReference class performed the same basic
functionality as the JNA PointerByReference class. Its main difference
was that it implemented a finalize() method which would free the
pointed-to string. Unfortunately, that finalize method contained errors
that could crash the JVM.

Replacement logic has been written that cleans up the pointed-to string
right away, so the LPWSTRByReference class is no longer needed.